re-organize components
